--- apps/builder/app/builder/features/sidebar-left/navigator/index.ts
@@ -1,1 +1,0 @@
-export * from "./navigator";
