+++ apps/builder/app/builder/features/sidebar-left/panels/assets/assets.tsx
@@ -1,18 +1,17 @@
 import { ImageIcon } from "@webstudio-is/icons";
-import { Flex } from "@webstudio-is/design-system";
 import { ImageManager } from "~/builder/shared/image-manager";
 import type { TabContentProps } from "../../types";
-import { Header, CloseButton } from "../../shared/header";
+import { Header, CloseButton, Root } from "../../shared/panel";
 
 export const TabContent = ({ onSetActiveTab }: TabContentProps) => {
   return (
-    <Flex direction="column" css={{ height: "100%" }}>
+    <Root>
       <Header
         title="Assets"
         suffix={<CloseButton onClick={() => onSetActiveTab("none")} />}
       />
       <ImageManager />
-    </Flex>
+    </Root>
   );
 };
