+++ apps/builder/app/builder/features/sidebar-left/panels/marketplace/about.tsx
@@ -7,7 +7,7 @@ import {
   theme,
 } from "@webstudio-is/design-system";
 import type { MarketplaceOverviewItem } from "~/shared/marketplace/types";
-import { Header } from "../../shared/header";
+import { Header } from "../../shared/panel";
 import { ChevronDoubleLeftIcon } from "@webstudio-is/icons";
 
 export const About = ({
