+++ apps/builder/app/builder/features/sidebar-left/panels/navigator/index.ts
@@ -1,1 +1,2 @@
 export * from "./navigator";
+export * from "./navigator-content";
